[#201] refactor: 미팅랜덤 컴포넌트 분리
