+++ src/pages/Meeting22/Styles.ts
@@ -21,26 +21,3 @@ export const Meeting22Title = styled.div`
 export const Meeting22Container = styled.div`
   height: 78.5%;
 `;
-
-export const TicketCount = styled.div<{ $isRandomLoading: boolean }>`
-  font-size: 12px;
-  color: ${({ $isRandomLoading }) => ($isRandomLoading ? '#FFFFFF' : '#000')};
-  align-self: center;
-  font-weight: bold;
-  margin-top: 1%;
-  background-color: ${({ $isRandomLoading }) => ($isRandomLoading ? '#FFFFFF' : '#F2F2F2')};
-  height: 4%;
-  width: 30%;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  border-radius: 20px;
-  margin-left: auto;
-  margin-right: auto;
-
-  @media (max-width: 768px) {
-  @supports (-webkit-touch-callout: none) {
-      width: 25%;
-    }
-  }
-`;
